+++ openscenario/openscenario_interpreter/src/syntax/relative_distance_condition.cpp
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <openscenario_interpreter/error.hpp>
 #include <openscenario_interpreter/reader/attribute.hpp>
 #include <openscenario_interpreter/syntax/entities.hpp>  // TEMPORARY (TODO REMOVE THIS LINE)
 #include <openscenario_interpreter/syntax/relative_distance_condition.hpp>
@@ -40,6 +41,12 @@ RelativeDistanceCondition::RelativeDistanceCondition(
   triggering_entities(triggering_entities),
   results(triggering_entities.entity_refs.size(), Double::nan())
 {
+  std::set<RoutingAlgorithm::value_type> supported = {
+    RoutingAlgorithm::value_type::shortest, RoutingAlgorithm::value_type::undefined};
+  if (supported.find(routing_algorithm) == supported.end()) {
+    throw UNSUPPORTED_ENUMERATION_VALUE_SPECIFIED(
+      RelativeDistanceCondition, boost::lexical_cast<std::string>(routing_algorithm));
+  }
 }
 
 auto RelativeDistanceCondition::description() const -> String